Refactor special commands into a package.
